requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
PyJWT==2.6.0
requests
psycopg2-binary
flask-restful
marshmallow_sqlalchemy
flask_jwt_extended